+++ jest.config.js
@@ -26,4 +26,5 @@ export default {
     ],
   },
   transformIgnorePatterns: ['/node_modules/'],
+  testPathIgnorePatterns: ['/node_modules/', '/dist/', '/frontend/'],
 };